Moved all non-gui stuff outside the GUI Added FindObject and ObjSignature classes Added ULogger from utilite for pretty log
git-svn-id: http://find-object.googlecode.com/svn/trunk/find_object@356 620bd6b2-0a58-f614-fd9a-1bd335dccda9

+++ src/rtabmap/PdfPlot.cpp
@@ -119,7 +119,7 @@ void PdfPlotCurve::clear()
 	UPlotCurve::clear();
 }
 
-void PdfPlotCurve::setData(const QMap<int, float> & dataMap, const QMap<int, int> & weightsMap)
+void PdfPlotCurve::setData(const QMap<int, int> & dataMap, const QMap<int, int> & weightsMap)
 {
 	ULOGGER_DEBUG("dataMap=%d, weightsMap=%d", dataMap.size(), weightsMap.size());
 	if(dataMap.size() > 0)
@@ -146,7 +146,7 @@ void PdfPlotCurve::setData(const QMap<int, float> & dataMap, const QMap<int, int
 		// update values
 		QList<QGraphicsItem*>::iterator iter = _items.begin();
 		QMap<int, int>::const_iterator j=weightsMap.begin();
-		for(QMap<int, float>::const_iterator i=dataMap.begin(); i!=dataMap.end(); ++i, ++j)
+		for(QMap<int, int>::const_iterator i=dataMap.begin(); i!=dataMap.end(); ++i, ++j)
 		{
 			((PdfPlotItem*)*iter)->setLikelihood(i.key(),  i.value(), j!=weightsMap.end()?j.value():-1);
 			//2 times...
